# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InquiryInputInquiryType = typeof InquiryInputInquiryType[keyof typeof InquiryInputInquiryType];


export const InquiryInputInquiryType = {
  AI_트레이딩: 'AI 트레이딩',
  IPO: 'IPO',
  채권: '채권',
  메자닌: '메자닌',
  기타: '기타',
} as const;

export interface InquiryInput {
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  /**
     * @minLength 8
     * @maxLength 30
     */
  phone: string;
  /** @maxLength 120 */
  email: string;
  inquiryType: InquiryInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}

export type Inquiry = InquiryInput & {
  id: number;
  createdAt: string;
};

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface PerformanceAiMonth {
  /**
     * @minLength 1
     * @maxLength 30
     */
  month: string;
  /**
     * @minimum -1000
     * @maximum 1000
     */
  return: number;
}

export interface PerformanceAiTrade {
  /**
     * @minLength 1
     * @maxLength 100
     */
  stockName: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  tradeDetail: string;
  /**
     * @minimum -1000
     * @maximum 1000
     */
  return: number;
}

export interface PerformanceAiInput {
  /**
     * @minItems 12
     * @maxItems 12
     */
  months: PerformanceAiMonth[];
  /**
     * @minItems 5
     * @maxItems 5
     */
  recentTrades: PerformanceAiTrade[];
  /**
     * @minimum -1000
     * @maximum 1000
     */
  annualAverageReturn: number;
  /**
     * @minimum 1
     * @maximum 100000
     */
  annualAverageStockCount: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  annualWinRate: number;
}

export type PerformanceAi = PerformanceAiInput & {
  updatedAt: string;
};

export interface PerformanceIpoInvestment {
  /**
     * @minLength 1
     * @maxLength 100
     */
  stockName: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  purchasePrice: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  purchasePeriod: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  listingDate: string;
  /**
     * @minimum -1000
     * @maximum 10000
     */
  return: number;
}

export interface PerformanceIpoInput {
  /**
     * @minItems 13
     * @maxItems 13
     */
  investments: PerformanceIpoInvestment[];
}

export type PerformanceIpo = PerformanceIpoInput & {
  updatedAt: string;
};

export interface PerformanceBondHolding {
  /**
     * @minLength 1
     * @maxLength 200
     */
  bondName: string;
  /**
     * @minimum 0
     * @maximum 100000
     */
  duration: number;
  /**
     * @minimum 0
     * @maximum 100000
     */
  weight: number;
}

export interface PerformanceBondsInput {
  /**
     * @minItems 10
     * @maxItems 10
     */
  holdings: PerformanceBondHolding[];
}

export type PerformanceBonds = PerformanceBondsInput & {
  updatedAt: string;
};

export interface PerformanceMezzanineHolding {
  /**
     * @minLength 1
     * @maxLength 100
     */
  company: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  type: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  investmentPeriod: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  exitPeriod: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  returnRate: string;
}

export interface PerformanceMezzanineInput {
  /**
     * @minItems 5
     * @maxItems 5
     */
  holdings: PerformanceMezzanineHolding[];
}

export type PerformanceMezzanine = PerformanceMezzanineInput & {
  updatedAt: string;
};