# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryInputInquiryType } from './inquiryInputInquiryType';

export interface InquiryInput {
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  /**
     * @minLength 8
     * @maxLength 30
     */
  phone: string;
  /** @maxLength 120 */
  email: string;
  inquiryType: InquiryInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}